% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xTalkQuant.R
\name{quantify_crossTalk_jaccard_index}
\alias{quantify_crossTalk_jaccard_index}
\title{A function to quantify Type-I cross-talk among pathways from enrichment analysis}
\usage{
quantify_crossTalk_jaccard_index(
  neighbourhood_th,
  string_PPI_score_th,
  pathway_a = c(),
  pathway_b = c()
)
}
\arguments{
\item{neighbourhood_th}{A number to threshold the PPI neighborhood while calculating the network proximity}

\item{string_PPI_score_th}{A number to threshold for filtering PPI confidence scores from String Database}

\item{pathway_a}{A set of strings which indicates gene symbols}

\item{pathway_b}{A set of strings which indicates gene symbols}
}
\value{
a vector of values: 1) the jaccard-index (as Type-I cross-talk score), 2) number of common/shared genes, 3) the common gene symbols (comma seperated)
}
\description{
A function to quantify Type-I cross-talk among pathways from enrichment analysis
}
\examples{
""
""
}
