% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xTalkQuant.R
\name{quantify_crossTalk_network_proximity}
\alias{quantify_crossTalk_network_proximity}
\title{A function to quantify Type-II cross-talk among pathways from enrichment analysis}
\usage{
quantify_crossTalk_network_proximity(
  neighbourhood_th,
  string_PPI_score_th,
  pathway_a = c(),
  pathway_b = c(),
  nPermute
)
}
\arguments{
\item{neighbourhood_th}{A number to threshold the PPI neighborhood while calculating the network proximity}

\item{string_PPI_score_th}{A number to threshold for filtering PPI confidence scores from String Database}

\item{pathway_a}{A set of strings which indicates gene symbols}

\item{pathway_b}{A set of strings which indicates gene symbols}

\item{nPermute}{A number to indicate how many times the permutations should be done for computing p-values for each Type-II cross-talks}
}
\value{
Network proximity score as the Type-II cross-talk value
}
\description{
A function to quantify Type-II cross-talk among pathways from enrichment analysis
}
\examples{
""
""
}
