% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xTalkPlot.R
\name{xTalkPlot.Type_I.NetworkView}
\alias{xTalkPlot.Type_I.NetworkView}
\title{A function to visualize (Network plot) Type-I cross-talk among pathways from enrichment analysis}
\usage{
xTalkPlot.Type_I.NetworkView(
  enrich.df = NULL,
  string_PPI_score_th = 900,
  showCategory = 100,
  layout = "kk",
  colorEdge = T,
  circular = F
)
}
\arguments{
\item{enrich.df}{Enrichment object (a R dataframe) that has the following column names: "Description", "Pathway.geneSymbol","Overlapping.geneSymbol","Overlapping.geneID", "Count", "GeneRatio", "pvalue","p.adjust")}

\item{string_PPI_score_th}{A number to threshold for filtering PPI confidence scores from String Database}

\item{showCategory}{A number for thresholding the number pathways to be included in the analysis}

\item{layout}{A string that should indicate the layout name (see igraph layout names)}

\item{colorEdge}{A logical value to decide if the edges should be colored in the network plots}

\item{circular}{A logical value to decide if the network plot should be circular}
}
\value{
A ggplot2 object
}
\description{
A function to visualize (Network plot) Type-I cross-talk among pathways from enrichment analysis
}
\examples{
library(rXTalkViz)
library(dplyr)
filtered_df <- example_enrich.df \%>\% filter(p.adjust < 0.001)
xTalkPlot.Type_I.NetworkView(enrich.df = filtered_df,
                            string_PPI_score_th = string_PPI_score_th,
                            showCategory = 100,
                            layout = "linear",
                            colorEdge = TRUE,
                            circular = TRUE)
}
