% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xTalkPlot.R
\name{xTalkPlot.Type_I.Sankey}
\alias{xTalkPlot.Type_I.Sankey}
\title{A function to visualize (Sankey plot) Type-I cross-talk among pathways from enrichment analysis}
\usage{
xTalkPlot.Type_I.Sankey(enrich.df = NULL, outdir = NULL)
}
\arguments{
\item{enrich.df}{Enrichment object (a R dataframe) that has the following column names: "Description", "Pathway.geneSymbol","Overlapping.geneSymbol","Overlapping.geneID", "Count", "GeneRatio", "pvalue","p.adjust")}

\item{outdir}{A directory that indicates where the outputs will be saved (all the plot PDFs, raw ggplot2 objects and the final report as csv)}
}
\value{
""
}
\description{
A function to visualize (Sankey plot) Type-I cross-talk among pathways from enrichment analysis
}
\examples{
""
}
