% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xTalkPlot.R
\name{xTalkPlot.Type_II.heatmap}
\alias{xTalkPlot.Type_II.heatmap}
\title{A function to visualize (Heatmap plot) Type-II cross-talk among pathways from enrichment analysis}
\usage{
xTalkPlot.Type_II.heatmap(df = NULL)
}
\arguments{
\item{df}{A R dataframe that has pair-wise cross-talk information (measured via network proximity scores)}
}
\value{
a heatmap object from ComplexHeatmap package
}
\description{
A function to visualize (Heatmap plot) Type-II cross-talk among pathways from enrichment analysis
}
\examples{
""
""
}
