% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xTalkPlot.R
\name{xTalkPlot.Type_II.NetworkView}
\alias{xTalkPlot.Type_II.NetworkView}
\title{A function to visualize (Network plot) Type-II cross-talk among pathways from enrichment analysis}
\usage{
xTalkPlot.Type_II.NetworkView(df = NULL, isCircular, layout)
}
\arguments{
\item{df}{A R dataframe that has pair-wise cross-talk information (measured via network proximity scores)}

\item{isCircular}{A logical value to decide if the network plot should be circular}

\item{layout}{A string that should indicate the layout name (see igraph layout names)}
}
\value{
A ggplot2 object
}
\description{
A function to visualize (Network plot) Type-II cross-talk among pathways from enrichment analysis
}
\examples{
""
""
}
